Prevent exposing internal mctp_binding_smbus struct from getting exposed
